spiderfoot: Add SpiderFootHelpers
