+++ spiderfoot/__init__.py
@@ -2,4 +2,5 @@ from .db import SpiderFootDb
 from .event import SpiderFootEvent
 from .plugin import SpiderFootPlugin
 from .target import SpiderFootTarget
+from .helpers import SpiderFootHelpers
 from spiderfoot.__version__ import __version__
